+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* Support for 'Oracle' GeoServer store/repository types [#71](https://gitlab.data.bas.ac.uk/MAGIC/web-map-inventory/issues/71)
+
 ## [0.3.1] - 2020-04-29
 
 ### Added